requirements.txt
absl-py==2.1.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
Cartopy==0.24.1
cdsapi==0.7.5
certifi==2024.12.14
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4.post1
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
cloudpickle==3.1.1
colabtools==0.0.1
comm==0.2.2
contourpy==1.3.1
crc32c==2.7.1
cycler==0.12.1
dask==2025.1.0
datapi==0.1.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dinosaur==1.2.1
dinosaur-dycore==1.2.1
dm-haiku==0.0.13
dm-tree==0.1.8
donfig==0.8.1.post1
eccodes==2.39.2
etils==1.11.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.17.0
findlibs==0.0.5
Flask==3.1.0
flatbuffers==25.1.24
flax==0.10.2
flexcache==0.3
flexparser==0.4
fonttools==4.55.4
fqdn==1.5.1
fsspec==2024.12.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphcast @ https://github.com/deepmind/graphcast/archive/master.zip#sha256=cab1b8a9aded275fdee7210f10066612fd75921b4e19ec0eac53c6141ecd3c2d
grpcio==1.70.0
h11==0.14.0
h5netcdf==1.5.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanize==4.11.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.0
jax-cuda12-pjrt==0.5.0
jax-cuda12-plugin==0.5.0
jaxlib==0.5.0
jaxline==0.0.8
jedi==0.19.2
Jinja2==3.1.5
jmp==0.0.4
joblib==1.4.2
jraph==0.0.6.dev0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.4.1
ml_collections==1.0.0
mpmath==1.3.0
msgpack==1.1.0
multiurl==0.3.3
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numcodecs==0.15.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.0
orbax-checkpoint==0.11.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.1.0
Pint==0.24.4
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
shapely==2.0.6
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorstore==0.1.71
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
tree-math==0.2.1
trimesh==4.6.0
triton==3.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.1.1
xarray-tensorstore==0.1.5
zarr==3.0.1
zipp==3.21.0
